requirements.txt
aiofiles==24.1.0
alembic==1.16.5
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
autogenstudio==0.4.2.2
azure-ai-agents==1.2.0b3
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-ai-projects==1.1.0b3
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
cryptography==45.0.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
flatbuffers==25.2.10
greenlet==3.2.4
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
loguru==0.7.3
lxml==6.0.1
magika==0.6.2
Mako==1.3.10
mammoth==1.10.0
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
numpy==1.26.4
olefile==0.47
onnxruntime==1.22.1
openai==1.102.0
openpyxl==3.1.5
opentelemetry-api==1.36.0
packaging==25.0
pandas==2.3.2
pyautogen==0.5.3
pdfminer.six==20250506
pillow==11.3.0
playwright==1.54.0
protobuf==5.29.5
psycopg==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.3
sympy==1.14.0
tiktoken==0.11.0
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xlrd==2.0.2
xlsxwriter==3.2.5
youtube-transcript-api==1.0.3
zipp==3.23.0
